Pull Joomla! download URLs
This retrieves the download URLs from GitHub.

Together with the manually collected list, it allows us to get the
download URLs of every published Joomla! version.

Note that there are THREE versions in Joomla's history which were only
available as an update, not a full installation download. As a result,
there will be no download information for them.

Thankfully, crap like that is a thing of the past. Phew.